+++ lib/pages/workout/workout_04_finish_page.dart
@@ -83,7 +83,7 @@ class WorkoutFinishPage extends StatelessWidget {
                     });
                   }
                 }
-              });
+              }).then((value) => wp.resetAllData());
               Navigator.pushAndRemoveUntil(
                   context,
                   PageTransition(