+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.1
 - Changed alarm sound.
+- Added "Stop the rest" button on timer screen.
 
 ## 1.0
 First version.