+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 1.1
 - Changed alarm sound.
 - Added "Stop the rest" button on timer screen.
-- Resetting workoud data after finish.
+- Resetting workout data after finish.
 
 ## 1.0
 First version.