+++ CHANGELOG.md
@@ -1,11 +1,9 @@
 ## 1.1
 - Changed alarm sound.
 - Added "Stop the rest" button on timer screen.
-- Resetting workout data after finish.
+- Fix: Resetting workout data after finish.
 - New Settings screen.
 - Share database.
-- Backup database.
-- Restore database.
 - Import database.
 
 ## 1.0