+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.1.1
+- Fixes and perfomance improvance.
+- Replaced "Edit" name buttons to longtap. 
+
 ## 1.1
 - Changed alarm sound.
 - Added "Stop the rest" button on timer screen.